+++ packages/3id-did-resolver/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](/compare/@ceramicnetwork/3id-did-resolver@2.0.4...@ceramicnetwork/3id-did-resolver@2.1.0) (2022-06-06)
+
+
+### Features
+
+* **3id-did-resolver:** throw error on failed multiquery (#2206) 68b1771, closes #2206
+
+
+
+
+
 ## [2.0.4](/compare/@ceramicnetwork/3id-did-resolver@2.0.4-rc.2...@ceramicnetwork/3id-did-resolver@2.0.4) (2022-05-18)
 
 **Note:** Version bump only for package @ceramicnetwork/3id-did-resolver