+++ packages/stream-model-instance/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](/compare/@ceramicnetwork/stream-model-instance@0.0.2...@ceramicnetwork/stream-model-instance@0.1.0) (2022-06-06)
+
+
+### Features
+
+* **stream-model-instance, stream-model-instance-handler:** ModelInstanceDocument API (#2196) 3ecf9fd, closes #2196
+
+
+
+
+
 ## [0.0.2](/compare/@ceramicnetwork/stream-model-instance@0.0.2-rc.4...@ceramicnetwork/stream-model-instance@0.0.2) (2022-05-18)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance